+++ grasshopperfund/organizations/models.py
@@ -30,20 +30,3 @@ class Organization(models.Model):
     @property
     def all_campaigns(self) -> list:
         return self.campaigns.all()
-
-
-class Post(models.Model):
-    # Foreign keys
-    author = models.ForeignKey(User, on_delete=models.CASCADE, related_name='posts')
-
-    # allows us to access an organization's posts
-    organization = models.ForeignKey(Organization, on_delete=models.CASCADE, related_name='posts')
-
-    # What should char limit be?
-    text = models.TextField(max_length=1000)
-
-    created = models.DateTimeField(auto_now=True)
-
-
-    def __str__(self):
-        return f"author: {self.author} \norganization: {self.organization} \ntext:{self.text}"
