+++ grasshopperfund/organizations/tests/test_models.py
@@ -3,7 +3,8 @@ import random
 from django.test import TestCase
 from django.contrib.auth.models import User
 
-from ..models import Organization, Post
+from ...posts.models import Post
+from ..models import Organization
 
 class TestModels(TestCase):
     def setUp(self):
