+++ grasshopperfund/posts/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from .models import Post
+
+# Register your models here.
+
+admin.site.register(Post)
