+++ grasshopperfund/posts/apps.py
@@ -1,0 +1,9 @@
+from django.apps import AppConfig
+
+
+
+class PostsConfig(AppConfig):
+    name = 'grasshopperfund.posts'
+
+    def ready(self):
+        pass
