+++ grasshopperfund/tags/models.py
@@ -1,6 +1,6 @@
 from django.db import models
 
-from ..organizations.models import Organization, Post
+from ..organizations.models import Organization
 
 # Create your models here.
 class Tags(models.Model):
