+++ grasshopperfund/templates/organizations/view_organization.html
@@ -1,4 +1,4 @@
-xtends 'main.html' %}
+{% extends 'main.html' %}
 {% load humanize %}
 {% block content %}
 {% load material_form %}
