+++ grasshopperfund/templates/posts/view_post.html
@@ -1,0 +1,23 @@
+{% extends 'main.html' %}
+{% load humanize %}
+{% block content %}
+{% load material_form %}
+
+<br>
+
+<div class="row">
+  <!-- each post uses materializecss card class -->
+  <div class="card darken-1 hoverable small">
+    <div class="card-content ">
+      <span class="card-title">{{post.author}}</span>
+
+      <p class="flow-text">
+        <ul>Post text: {{post.text}}</ul>
+        <ul>{{post.created}}</ul>
+      </p>
+    </div>
+  </div>
+
+
+
+</div>
